Customer Churn Prediction
Predicting the churn for bank customers where we have built classification model , in this dataset exited customers is the target variable and other are independent variables. In this data prediction we have used logistic regression , random forest and xgboost classifier to make predictions better and comparable with each other and choose the right one for this data in making future goals great, here, we have selected the xgboost classification model because it is throwing a good accuracy and values as compare to others.